task(SDK-5584) - Refactors Media Handling for InApps
