+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/CTInAppNotification.kt
@@ -286,10 +286,6 @@ class CTInAppNotification : Parcelable {
         dest.writeByte((if (isRequestForPushPermission) 0x01 else 0x00).toByte())
     }
 
-    fun hasStreamMedia(): Boolean {
-        return !_mediaList.isEmpty() && _mediaList[0].isMediaStreamable()
-    }
-
     internal fun getInAppMediaForOrientation(orientation: Int): CTInAppNotificationMedia? {
         var returningMedia: CTInAppNotificationMedia? = null
         for (inAppNotificationMedia in _mediaList) {
@@ -344,7 +340,7 @@ class CTInAppNotification : Parcelable {
         try {
             id = jsonObject.optString(Constants.INAPP_ID_IN_PAYLOAD, "")
             campaignId = jsonObject.optString(Constants.NOTIFICATION_ID_TAG, "")
-            type = jsonObject.getString(Constants.KEY_TYPE) // won't be null based on constructor
+            type = "cover-image" // won't be null based on constructor
             isLocalInApp = jsonObject.optBoolean(CTLocalInApp.Companion.IS_LOCAL_INAPP, false)
             fallBackToNotificationSettings = jsonObject.optBoolean(
                 CTLocalInApp.Companion.FALLBACK_TO_NOTIFICATION_SETTINGS, false
