+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/fragment/CTInAppNativeHeaderFragment.kt
@@ -12,11 +12,11 @@ import android.widget.LinearLayout
 import android.widget.RelativeLayout
 import android.widget.TextView
 import androidx.core.graphics.toColorInt
-import androidx.media3.common.util.UnstableApi
 import com.clevertap.android.sdk.R
+import com.clevertap.android.sdk.inapp.media.InAppMediaConfig
+import com.clevertap.android.sdk.inapp.media.InAppMediaDelegate
 import com.clevertap.android.sdk.applyInsetsWithMarginAdjustment
 
-@UnstableApi
 internal class CTInAppNativeHeaderFragment : CTInAppBasePartialNativeFragment() {
 
     private lateinit var mediaDelegate: InAppMediaDelegate
@@ -108,6 +108,11 @@ internal class CTInAppNativeHeaderFragment : CTInAppBasePartialNativeFragment() 
     override fun onStart() {
         super.onStart()
         mediaDelegate.onStart()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        mediaDelegate.onResume()
     }
 
     override fun onPause() {
