+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/fragment/CTInAppNativeInterstitialFragment.kt
@@ -15,11 +15,11 @@ import android.widget.LinearLayout
 import android.widget.RelativeLayout
 import android.widget.TextView
 import androidx.core.graphics.toColorInt
-import androidx.media3.common.util.UnstableApi
 import com.clevertap.android.sdk.R
+import com.clevertap.android.sdk.inapp.media.InAppMediaConfig
+import com.clevertap.android.sdk.inapp.media.InAppMediaDelegate
 import com.clevertap.android.sdk.customviews.CloseImageView
 
-@UnstableApi
 internal class CTInAppNativeInterstitialFragment : CTInAppBaseFullNativeFragment() {
 
     private lateinit var mediaDelegate: InAppMediaDelegate
@@ -32,9 +32,9 @@ internal class CTInAppNativeInterstitialFragment : CTInAppBaseFullNativeFragment
             inAppNotification = inAppNotification,
             currentOrientation = currentOrientation,
             isTablet = inAppNotification.isTablet && isTablet(),
-            resourceProvider = resourceProvider()
+            resourceProvider = resourceProvider(),
+            supportsStreamMedia = true
         )
-        mediaDelegate.initVideoPlayerHandle()
     }
 
     @SuppressLint("ResourceType")
@@ -54,8 +54,6 @@ internal class CTInAppNativeInterstitialFragment : CTInAppBaseFullNativeFragment
         closeImageView = fl.findViewById(CloseImageView.VIEW_ID)
         relativeLayout = fl.findViewById(R.id.interstitial_relative_layout)
 
-        mediaDelegate.bindVideoFrame(relativeLayout?.findViewById(R.id.video_frame))
-
         // Container backgrounds
         relativeLayout?.setBackgroundColor(inAppNotification.backgroundColor.toColorInt())
         fl.background = ColorDrawable(0xBB000000.toInt())
@@ -71,6 +69,7 @@ internal class CTInAppNativeInterstitialFragment : CTInAppBaseFullNativeFragment
                 clickableMedia = false,
                 useOrientationForImage = false,
                 hideImageViewForNonImageMedia = false,
+                videoFrameId = R.id.video_frame,
                 fillVideoFrame = false
             )
         )
@@ -114,7 +113,7 @@ internal class CTInAppNativeInterstitialFragment : CTInAppBaseFullNativeFragment
             closeImageView?.setVisibility(View.VISIBLE)
             closeImageView?.setOnClickListener {
                 didDismiss(null)
-                mediaDelegate.clearGif()
+                mediaDelegate.clear()
                 activity?.finish()
             }
         }
