+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/fragment/CTInAppNativeInterstitialImageFragment.kt
@@ -10,11 +10,11 @@ import android.view.ViewTreeObserver.OnGlobalLayoutListener
 import android.widget.FrameLayout
 import android.widget.RelativeLayout
 import androidx.core.graphics.toColorInt
-import androidx.media3.common.util.UnstableApi
 import com.clevertap.android.sdk.R
+import com.clevertap.android.sdk.inapp.media.InAppMediaConfig
+import com.clevertap.android.sdk.inapp.media.InAppMediaDelegate
 import com.clevertap.android.sdk.customviews.CloseImageView
 
-@UnstableApi
 internal class CTInAppNativeInterstitialImageFragment : CTInAppBaseFullFragment() {
 
     private lateinit var mediaDelegate: InAppMediaDelegate
@@ -27,9 +27,9 @@ internal class CTInAppNativeInterstitialImageFragment : CTInAppBaseFullFragment(
             inAppNotification = inAppNotification,
             currentOrientation = currentOrientation,
             isTablet = inAppNotification.isTablet && isTablet(),
-            resourceProvider = resourceProvider()
+            resourceProvider = resourceProvider(),
+            supportsStreamMedia = true
         )
-        mediaDelegate.initVideoPlayerHandle()
     }
 
     override fun onCreateView(
@@ -49,8 +49,6 @@ internal class CTInAppNativeInterstitialImageFragment : CTInAppBaseFullFragment(
         relativeLayout = fl.findViewById(R.id.interstitial_image_relative_layout)
 
         relativeLayout?.setBackgroundColor(inAppNotification.backgroundColor.toColorInt())
-
-        mediaDelegate.bindVideoFrame(relativeLayout?.findViewById(R.id.video_frame))
 
         when (currentOrientation) {
             Configuration.ORIENTATION_PORTRAIT -> relativeLayout?.getViewTreeObserver()
@@ -120,13 +118,13 @@ internal class CTInAppNativeInterstitialImageFragment : CTInAppBaseFullFragment(
 
         mediaDelegate.setMediaForInApp(
             relativeLayout,
-            InAppMediaConfig(imageViewId = R.id.interstitial_image, clickableMedia = true),
+            InAppMediaConfig(imageViewId = R.id.interstitial_image, clickableMedia = true, videoFrameId = R.id.video_frame),
             CTInAppNativeButtonClickListener()
         )
 
         closeImageView.setOnClickListener {
             didDismiss(null)
-            mediaDelegate.clearGif()
+            mediaDelegate.clear()
             activity?.finish()
         }
